+++ datacontract/cli.py
@@ -159,6 +159,7 @@ def export(
 
 class ImportFormat(str, Enum):
     sql = "sql"
+    avro = "avro"
 
 
 @app.command(name="import")